+++ UI/__init__.py
@@ -1,0 +1,23 @@
+from .main_window import MultiChatWindow
+from .components import (
+    StyledButton,
+    StyledComboBox,
+    StyledCheckBox,
+    StyledTextEdit,
+    StyledPlainTextEdit,
+    ChatPanel,
+    LoadingIndicator,
+    TemplateDialog
+)
+
+__all__ = [
+    'MultiChatWindow',
+    'StyledButton',
+    'StyledComboBox',
+    'StyledCheckBox',
+    'StyledTextEdit',
+    'StyledPlainTextEdit',
+    'ChatPanel',
+    'LoadingIndicator',
+    'TemplateDialog'
+]
