+++ UI/components/__init__.py
@@ -1,0 +1,21 @@
+from .styled_widgets import (
+    StyledButton,
+    StyledComboBox,
+    StyledCheckBox,
+    StyledTextEdit,
+    StyledPlainTextEdit
+)
+from .chat_panel import ChatPanel
+from .loading_indicator import LoadingIndicator
+from .template_dialog import TemplateDialog
+
+__all__ = [
+    'StyledButton',
+    'StyledComboBox',
+    'StyledCheckBox',
+    'StyledTextEdit',
+    'StyledPlainTextEdit',
+    'ChatPanel',
+    'LoadingIndicator',
+    'TemplateDialog'
+]
