+++ __init__.py
@@ -1,0 +1,13 @@
+from .config import Settings, ConfigManager
+from .services import APIService, HistoryService
+from .UI import MultiChatWindow
+from .utils import ProxyUtils
+
+__all__ = [
+    'Settings',
+    'ConfigManager',
+    'APIService',
+    'HistoryService',
+    'MultiChatWindow',
+    'ProxyUtils'
+]
