+++ utils/__init__.py
@@ -1,0 +1,3 @@
+from .proxy_utils import ProxyUtils
+
+__all__ = ['ProxyUtils']
